+++ web/src/components/molecules/Asset/InfoBox/index.tsx
